fix: rename external guarantor references to improve clarity

+++ contracts/0.8.25/vaults/Permissions.sol
@@ -263,7 +263,7 @@ abstract contract Permissions is AccessControlConfirmable {
     function _withdrawDisprovenValidatorFromPDG(
         bytes calldata _pubkey,
         address _recipient
-    ) internal onlyRole(PDG_WITHDRAWAL_ROLE) returns (uint128) {
+    ) internal onlyRole(PDG_WITHDRAWAL_ROLE) returns (uint256) {
         return PredepositGuarantee(stakingVault().depositor()).withdrawDisprovenPredeposit(_pubkey, _recipient);
     }
